// src/modules/vaccine-export-detail/types/vaccine-export-detail.type.ts
import type { Disease } from "@/modules/diseases";
import type { Vet } from "@/modules/vets/types/vet.type";

export interface VaccineExportDetail {
  vaccineExportDetailId: number | null;
  vaccineBatchId: number | null;
  vaccineExportId: number | null;
  appointmentDetailId: number | null;
  quantity: number;
  purpose: string;
  notes: string;
  createdAt: string;
  createdBy: string;
  modifiedAt: string;
  modifiedBy: string;
  isDeleted: boolean;
  vaccineBatch: {
    vaccineBatchId: number | null;
    vaccineId: number | null;
    batchNumber: string;
    manufactureDate: string;
    expiryDate: string;
    manufacturer: string;
    source: string;
    storageCondition: null;
    quantity: number;
    createdAt: string;
    createdBy: string;
    modifiedAt: string;
    modifiedBy: string;
    vaccineResponseDTO: {
      vaccineId: number | null;
      vaccineCode: string;
      name: string;
      description: string;
      price: number;
      status: string;
      image: string;
      notes: string;
      createdAt: string;
      createdBy: string;
      modifiedAt: string;
      modifiedBy: string;
      isDeleted: false;
    };
  };
  vaccineExport: {
    vaccineExportId: number | null;
    exportCode: string;
    exportDate: string;
    createdAt: string;
    createdBy: string;
    modifiedAt: string;
    modifiedBy: string;
    isDeleted: boolean;
  };
  appointmentDetail: {
    appointmentDetailId: number | null;
    appointmentId: number | null;
    vetId: number | null;
    serviceType: number;
    microchipItemId: number | null;
    passportId: number | null;
    healthConditionId: number | null;
    vaccineBatchId: number | null;
    diseaseId: number | null;
    appointmentDetailCode: string;
    dose: string;
    reaction: string;
    nextVaccinationInfo: string;
    appointmentDate: string;
    appointmentStatus: number;
    temperature: string;
    heartRate: string;
    generalCondition: string;
    others: string;
    notes: null;
    createdAt: string;
    createdBy: string;
    modifiedAt: string;
    modifiedBy: string;
    isDeleted: boolean;
    vet: Vet;
    microchipItem: null;
    petPassport: null;
    healthCondition: null;
    vaccineBatch: {
      vaccineBatchId: number | null;
      vaccineId: number | null;
      batchNumber: string;
      manufactureDate: string;
      expiryDate: string;
      manufacturer: string;
      source: string;
      storageCondition: null;
      quantity: number;
      createdAt: string;
      createdBy: string;
      modifiedAt: string;
      modifiedBy: string;
      vaccineResponseDTO: {
        vaccineId: number | null;
        vaccineCode: string;
        name: string;
        description: string;
        price: number;
        status: string;
        image: string;
        notes: string;
        createdAt: string;
        createdBy: string;
        modifiedAt: string;
        modifiedBy: string;
        isDeleted: boolean;
      };
    };
    disease: Disease;
  };
}